+++ Source/Engine/Audio/AudioSFML.hpp
@@ -36,7 +36,7 @@ namespace water
             /// <param name="soundLimit"> The total number of sounds available at one time. </param>
             /// <param name="bgmMixer"> The mixer volume for background music, 0 to 1. </param>
             /// <param name="sfxMixer"> The mixer volume for sound effects, 0 to 1. </param>
-            void initialise (const size_t soundLimit, const float bgmMixer, const float sfxMixer) override final;
+            void initialise (const unsigned int soundLimit, const float bgmMixer, const float sfxMixer) override final;
 
             /// <summary> Doesn't actually do anything. </summary>
             void update() override final;
